+++ src/StardustSandbox/UI/Elements/SliceImage.cs
@@ -163,7 +163,7 @@ namespace StardustSandbox.UI.Elements
                 return;
             }
 
-            for (int i = 0; i < this.slices.Length; i++)
+            for (byte i = 0; i < this.slices.Length; i++)
             {
                 spriteBatch.Draw(this.Texture, this.slices[i].Position, this.slices[i].SourceRectangle, this.Color, 0f, Vector2.Zero, this.slices[i].Scale, SpriteEffects.None, 0f);
             }
